removed redundant clear_seq_and_finish_examples
