+++ vowpalwabbit/cb_adf.cc
@@ -295,15 +295,6 @@ void output_example_seq(vw& all, cb_adf& data, multi_ex& ec_seq)
   }
 }
 
-void clear_seq_and_finish_examples(vw& all, multi_ex& ec_seq)
-{
-  if (ec_seq.size() > 0)
-    for (example* ecc : ec_seq)
-      if (ecc->in_use)
-        VW::finish_example(all, ecc);
-  ec_seq.erase();
-}
-
 void finish_multiline_example(vw& all, cb_adf& data, multi_ex& ec_seq)
 {
   if (ec_seq.size() > 0)
