+++ vowpalwabbit/explore_eval.cc
@@ -5,6 +5,7 @@
 #include "cb_adf.h"
 #include "cb_explore_adf.h"
 #include "rand48.h"
+#include "gen_cs_example.h"
 
 //Do evaluation of nonstationary policies.
 //input = contextual bandit label
@@ -107,16 +108,6 @@ void output_example_seq(vw& all, explore_eval& data, multi_ex& ec_seq)
   }
 }
 
-
-void clear_seq_and_finish_examples(vw& all, multi_ex& ec_seq)
-{
-  if (ec_seq.size() > 0)
-    for (example* ecc : ec_seq)
-      if (ecc->in_use)
-        VW::finish_example(all, ecc);
-  ec_seq.erase();
-}
-
 void finish_multiline_example(vw& all, explore_eval& data, multi_ex& ec_seq)
 {
   if (ec_seq.size() > 0)
@@ -124,7 +115,7 @@ void finish_multiline_example(vw& all, explore_eval& data, multi_ex& ec_seq)
     output_example_seq(all, data, ec_seq);
     CB_ADF::global_print_newline(all);
   }
-  clear_seq_and_finish_examples(all, ec_seq);
+  GEN_CS::clear_seq_and_finish_examples(all, ec_seq);
 }
 
 template <bool is_learn> void do_actual_learning(explore_eval& data, multi_learner& base, multi_ex& ec_seq)
